Add a dummy project for running the tests.
The tests we have right now weren't really designed as integration
tests, able to confirm that django-security is working correctly in
whatever project happens to have included it (we should probably write
some of those?) - they're unit tests, and they may trip up in a project
with the wrong configuration. So we've now placed them in a dummy
project whose settings.py we control.